+++ src/__init__.py
@@ -1,1 +1,18 @@
 """Package: prescription-trend-analyzer"""
+
+from src.anomaly_detector import AnomalyDetectorError, detect_anomalies
+from src.changepoint_detector import (
+    ChangePointDetectorError,
+    ChangePointResult,
+    detect_change_point,
+    detect_change_points,
+)
+
+__all__ = [
+    "AnomalyDetectorError",
+    "ChangePointDetectorError",
+    "ChangePointResult",
+    "detect_anomalies",
+    "detect_change_point",
+    "detect_change_points",
+]